refactor(utility): mempool (legacy and new)
